requirements.txt
absl-py==2.1.0
amqp==5.2.0
annotated-types==0.7.0
anyio==4.4.0
astunparse==1.6.3
billiard==4.2.0
celery==5.4.0
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
dnspython==2.6.1
email_validator==2.2.0
fastapi==0.112.1
flatbuffers==24.3.25
gast==0.6.0
google-pasta==0.2.0
grpcio==1.66.0
h11==0.14.0
h5py==3.11.0
idna==3.8
joblib==1.4.2
keras==3.5.0
kombu==5.4.0
libclang==18.1.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
ml-dtypes==0.4.0
motor==3.5.1
namex==0.0.8
numpy==1.26.4
opt-einsum==3.3.0
optree==0.12.1
packaging==24.1
pandas==2.2.2
passlib==1.7.4
prompt_toolkit==3.0.47
protobuf==4.25.4
pydantic==2.8.2
pydantic_core==2.20.1
Pygments==2.18.0
PyJWT==2.9.0
pymongo==4.8.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2024.1
requests==2.32.3
rich==13.7.1
setuptools==73.0.1
six==1.16.0
sniffio==1.3.1
starlette==0.38.2
tensorboard==2.17.1
tensorboard-data-server==0.7.2
tensorflow==2.17.0
termcolor==2.4.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
uvicorn==0.30.6
vine==5.1.0
wcwidth==0.2.13
Werkzeug==3.0.4
wheel==0.44.0
wrapt==1.16.0
xasm==1.2.0
